+++ src/utils/mod.rs
@@ -5,3 +5,4 @@ pub (crate) mod string_utils;
 pub (crate) mod json_utils;
 pub (crate) mod config_reader;
 pub (crate) mod multi_file_reader;
+pub (crate) mod file_reader;